backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
Pillow==9.5.0
pyttsx3==2.90
pytesseract==0.3.10
python-docx==0.8.11
